Implement fold-left.
This change introduces the Arguments interface, which is arguably a
better way to pass arguments to functions. When a function needs an
argument, it uses one of the functions defined in Arguments to get
the appropriate TypedObject. Any implementation details, such as
whether we need to evaluate an expression in order to provide the
TypedObject, are hidden from functions.