Introduced assignments and arrow functions
This change introduces a couple of syntactix sugars, namely the
assignment operator (symbol `:` expression) which is equivalent to
(define symbol expression), and arrow functions, e.g. (x) => (+ x 1)
which is equivalent to (lambda (x) (+ x 1)).

In order to remove a grammar clash between `=` and `=>`, the
equals operator has been renamed to `==` (previously it was `=`).